Add a data structure to hold the Env with support for case-insensitivity
github.com/buildkite/interpolate has a similar data structure we could
use, but it's case-sensitivity is not configurable and is determined by
the OS the code is executed on. It's a lot simpler if we could test both
cases on the same OS, and just supply the OS in the production code.