Log public key fingerprint while in debug
If the key id matches, but the public key isn't derived from the same private key that signed a step, then it's pretty hard to figure out why the signature verification fails.

We don't currently include the public key as part of the signature, just the key-id, so it's hard to validate they are the same.

This allows us to emit them to logs for easier comparison.