Pass ctx idiomatically
It shouldn't be passed through an option pattern, since that causes it to be stored in a struct. It should be passed as the first arg. See https://pkg.go.dev/context.

This adds a ctx arg to Verify, which uses it, and Sign, which doesn't, but now looks weird without it.